requirements.txt
pdf2image
pdfminer.six
tqdm
openai
pypdf
python-dotenv
